fix: clippy errors in tests in the compiler-core project
