+++ compiler-core/src/manifest.rs
@@ -267,6 +267,7 @@ where
 }
 
 #[cfg(test)]
+#[allow(clippy::unwrap_used)]
 mod tests {
     use super::*;
